gui/requirements.txt
wxpython
pyserial
